Move FileOptionType to formattedcode #2233
This Click option extension is only used there
