+++ src/formattedcode/__init__.py
@@ -21,3 +21,30 @@
 #  for any legal advice.
 #  ScanCode is a free software code scanning tool from nexB Inc. and others.
 #  Visit https://github.com/nexB/scancode-toolkit/ for support and download.
+
+from __future__ import absolute_import
+from __future__ import print_function
+from __future__ import unicode_literals
+
+from itertools import chain
+
+import click
+
+
+class FileOptionType(click.File):
+    """
+    A click.File subclass that ensures that a file name is not set to an
+    existing option parameter to avoid mistakes.
+    """
+
+    def convert(self, value, param, ctx):
+        known_opts = set(chain.from_iterable(
+            p.opts for p in ctx.command.params if isinstance(p, click.Option)))
+        if value in known_opts:
+            self.fail(
+                'Illegal file name conflicting with an option name: %s. '
+                'Use the special "-" file name to print results on screen/stdout.'
+                % (click.types.filename_to_ui(value),
+            ), param, ctx)
+        return click.File.convert(self, value, param, ctx)
+
