+++ src/formattedcode/output_jsonlines.py
@@ -31,10 +31,10 @@ import simplejson
 
 from commoncode.system import py2
 from commoncode.system import py3
+from formattedcode import FileOptionType
 from plugincode.output import output_impl
 from plugincode.output import OutputPlugin
 from scancode import PluggableCommandLineOption
-from scancode import FileOptionType
 from scancode import OUTPUT_GROUP
 
 """
